Move the test JSON to the test suite
The JSON we had here was a great test case.
We make it a legit case by moving it to the test suite.

This also means we need to make up a real `Page _` value.
We do that, and make a changes in the process.
The `key` fields were all the same initially.
It seems like we want them to be unique per `Field _`.