Setup CircleCI
CI can help us gain a bit more confidence when making our changes.
If the tests are always running when we push a branch or open aPR,
it should catch things that we humans miss from time to time.

Most of this is copied from `purescript-ocelot`.
If there's something we inherited that we shouldn't have,
we ought to change that and make it work better for us.