Start experimenting with `Lookup`
We're hitting a snag.

The lookup can return a value with some output type `o`.
We want the consumer to tell us how to transform a `Key` to an `o`.
We can take a function for that in `evalExpr'`,
but we need to have a different `o` than the `o` we're returning.
So we end up with something like:
```PureScript
evalExpr'
  :: ∀ o
   . Expressible o =>
   (forall o2. Key -> Maybe o2) ->
   Expr o ->
   o
```

That works for `evalExpr'`, but the consumer cannot decide what `o2` is,
since Rank-2 polymorphism doesn't work that way.
In the form, we're broken if we want to have a `Map Key o` for any `o`.

An alternative is to move the lookup function into `ExprF _ _`.
Unfortunately, we cannot encode PureScript functions into JSON.
We could push on that idea and add functions to `ExprF _ _`.
Then we could encode them into JSON (as they'd be plain old data).