Get things compiling again
This doesn't actually give us dynamic behavior.
The lookup is always the same.

We tried changing the monomorphized type to `ExprType`,
but it won't allow us to use the lookup to build a `Page Expr`.
Things like `description` have type `Expr String`,
and we'd need a function `Expr ExprType -> Expr String`.

Given the amount of trouble we're having, maybe our model is wrong?
We might want to go back and think about how to get what we want.
Restarting might be the quicker way forward in the long run.