Try a more dynamically typed approch
The well typed approach seems to be giving us some problems.
When we want to add lookup, we start running into typed issues.
Surely there's a way around it.

In the meantime, we explore an approach with more relaxed types.
We don't statically verify any of the expressions are well formed.
We dynamically check whether they can work.

This isn't the most ideal approach,
as we now have to deal with possible failures in multiple places.
However, it's forward progress.
We can also move to a more well-typed approach incrementally.

There are still ways we can make this better.
We might be able to type check at the edges,
and construct a better typed expression.
We might transform the failures into usable output.
We might do something else.

In any case, we now have some dynamic behavior in these forms!