Turn warnings into errors
As with a previous commit,
if we want to get lots of signal from the compiler,
we should mitigate the amount of noise we let through.

If we turn the warnings into errors, CI will catch these and fail on us.
We can address them quicker with a short feedback loop like this
and keep the compiler telling us useful things.