Use an `Expr` as the default instead of `ExprType`
There's no reason we have to restrict the default value to `ExprType`.
If someone wants to use an `Expr` as a default value, they can.

There was a concern that evaluation wouldn't terminate,
but `Expr` is inductive data.
Evaluation will terminate at some point.

This allows for a bit of choice when constructing lookups.
For instance:
```PureScript
lookup_ "foo" (lookup_ "bar" (lookup_ "baz" (boolean_ true)))
```

We check for three different keys in order,
and finally default to `true`.
Everything is still total, and all works well.