Switch over to a `Makefile`
Make tends to handle building of a project a bit better than
npm scripts as it grows in complexity.

We were just starting to get to that point.
We wanted to use `psa` to turn warnings into errors,
but `pulp` will mess up if we try to pass flags to `psa` all the time.
So we had to split it up into two scripts.
And we could only use the one script after the other had run.

There is a bit of work hiding in npm scripts though,
we have to always have things installed an up to date.
When first starting on the project, we have to remember to:
1. `npm install`
2. `npx bower install`
3. `npm run build`

When coming back to the project,
it's very possible the dependencies will change
and will require the same set of commands.

Or maybe the commands to get up to date have changed,
and we're using `psc-package` or `spago` or something.

Switching to a `Makefile` should be less painful in the long run.
Hopefully, people can come in, type `make` and be good to go.
No need to remember what order of commands to run,
no need to remember what commands to run, just `make`.