Remove some cases from values
Looks like we had 14 different cases to handle for rendering.
Most of these weren't cases that existed in our mental model.

Changing the types around a bit,
we were able to drop the cases that didn't make sense.

This should also make things a little easier for users to create.