Use `watch-exec` instead of `pulp --watch`
We run into an issue with `pulp --watch`:
on some machines, it will double build when a file changes.
Some of us have been using `watch-exec` instead,
and it's not had the same issue as `pulp --watch`.

We change the `Makefile` to use `watch-exec`,
so everyone is on the same page.