Drop the `Maybe _` around `InputSource _`
We were using `Maybe (InputSource _)` to represent three things:
* No value is set
* The value is set by the user
* The value is invalid
There were two conflicting ideas here.

One is that if the value is set (whether valid or invalid),
we were lumping them into the "successful" part of the `Maybe _`.
If we're looking to tell whether a required field will validate,
we have to check both whether it exists and whether it's not invalid.

The other is that if the field is not required and we want to tell
whether it will validate,
we have to check either that the value doesn't exist or the value
is not invalid.

Using something like:
```PureScript
data ValueError a
  = NotSet
  | Invalid a

type Value a = Either (ValueError a) a
```
would address the first case, but not the second.

Another shifting of `NotSet` to the "successful" part of `Either _ _`
would address the second case, but not the first.

Rather than bring in the connotations of `Maybe _` or `Either _ _`,
we give it a name relevant to our domain.
Hopefully, this should be clearer to everyone involed.

It does mean that we'll have to write explicit functions for validation,
but that doesn't seem like the worst thing (given our options).