Move auxiliary values to `Lynx.Data.Form`
These values we defined didn't need to live in the component.
They only traverse the `Page _` structure.

We can make it a bit clearer where responsibilities are
if the values we create are in the correct place.

An additional benefit of moving this stuff around is that
some of the values got simpler.
For instance, we don't need to pass around the `get` when `eval`ing.
We can nest the transforming values under the top-level `eval`
and let it go from there.
We could have done the same in `Lynx.Page.Form`, but we didn't.

This is pre-work for adding dropdown support.
We don't really want to go back and forth between modules
when fleshing that out.