Start implementing dropdown
This is the first stab at things.
It displays the dropdown, but it doesn't seem like it's all there yet.

For one, the `valueRepresentation` doesn't seem quite right.
Is it supposed to be a function?
If it is, how do we encode that to JSON?
Do we need to add functions to our `Expr`?

We're also ignoring two of the messages from the dropdown.
It seems like the `VisibilityChanged` message for sure should be handled.

Is there stuff we should be doing for `Emit`?
We're using `Const Void` as our `o` type. So, there's not much we can do.
Should we be using `Query`? That seems to happen in a few other places.
If so, we'll have something to do.

Are we rendering properly?
It feels like this was too easy to render.
So, we've probably forgotten something.