Don't set visibility based on the dropdown
We were wrong with what this was for.
It's some dropdown specific behavior.
When it's visible, the options in the dropdown are shown.
When it's not visible, the options in the dropdown are not shown.

That's useful to have,
but We don't want that behavior to change the visibility of the field.

PS-2958 #time 7m