PS-2958 Add Arrays and Pairs to `ExprType`
We want the `options` dropdowns to be any old `Expr` since there are cases
where we'll want to alter the `options` based on something else
in the `Page _`.
In order to support `Expr`s as `options`,
we'll need to have Arrays and key value pairs to encode `options`.

The `print` function is less clear what it's supposed to be doing,
but maybe it can be better named?