Remove `index` and `valueRepresentation`
These both ended up being unused.

The `index` is duplicating the `key` information.
The `key`s already have to be unique as we're using them for the slots.

The `valueRepresentation` is duplicating the name of each option.
If we decide to allow free-form options in the future,
we can bring something in there to convert from an option
to something we can render.