PS-2958 Render pairs more correctly
We were using `print` improperly.
It's supposed to be used for pretty printing values.

Rather than continue to abuse it for rendering `options`,
we do some projecting from `ExprType` and hope for the best.

We've renamed the `key` to `name`
and we're going to assume the `options` are the correct shape.
This will work so long as everything is well-typed.