PS-2958 Fix stack overflow
This is a problem caused by the polymoprhism of `arbitrary`.
If we were using explicit generators (rather than a typeclasses),
the compiler would have statically analyzed this line
and seen a type mismatch.
Since `arbitrary :: forall a. Arbitrary a => Gen a`,
it could find the `arbiraryExprType` value, and use it for solving.

We should move to generators and hopefully not run into this again.