PS-2958 Thread down the `options` changes
We have to do this manually in order to address:
https://github.com/citizennet/purescript-ocelot/issues/79.
The changes here allow the `options` to actually be dynamic.

Unfortunately, we have to remove the tests on the queries because of that.
There's no easy way to evaluate a `HalogenM _ _ _ _ _ _ _`.

Maybe someone will release an easy way to do this in the future,
and we can use that.
Until then, we're testless for our `eval`.