PS-2957 Move all the `ExprType` de/encoding together
We had the logic split between `Expr` and `ExprType` before.
This meant that we had to hop around to different places
when we wanted to do any de/encoding.

We can colocate this behavior so it's a bit easier to work with.

This becomes especially useful when we need to alter `ExprType`.
Adding another case to `ExprType`, for instance, can be unfun.