PS-2957 Add support for `Currency`
The previous couple of commits were in service of this one.
Make the change easy, then make the easy change, they say.
It should be noted that almost all of the changes in here are additive.
I think that means we're stumbling upon a decent code organization.

We needed to change to using `browserify` because we're now
(transitively) dependent on an `npm` package.
The previous changes to the `Makefile` made this a one-line change.

The rest of the work here is pretty straight forward.
The currency field does just enough to be useful.