PS-2961 Change the input to an ADT
We're about to have more than one route to deal with.
Since we're hardcoding data here,
we're going to hardcode the routes as well.

This isn't the final goal of wha the input is meant to represent,
While we're shooting for MVP, it should help us develop a little quicker.
E.g.: no need to explicilty parse strings, or forget cases or whatever.