PS-2961 Put together the MVP page
We're looking to take over the Campaign Request form.
So we're going to try to emulate that as much as possible.

For the first version of MVP,
we are looking to have just the name, objective, and media budget.
We finaly have all the support for these fields,
so we can throw them together in the form.