Pass down RTS flags to `purs`
What happens on Circle is that the machines might have 32 "physical" cores,
but only make two cores available to our build.
There's some issue where GHC based builds pick up the "physical" cores,
and use that as the number of threads to run.
If we rely on that behavior,
we get thrashing and the CPU ends up spending more time scheduling all
the parallelism than it does doing useful work.

Since `purs` is a GHC compiled binary with the right flags enabled,
we can adjust the RTS options of `purs`:
https://downloads.haskell.org/~ghc/8.4.3/docs/html/users_guide/runtime_control.html#setting-rts-options-on-the-command-line.

In this case, we tell it to use exactly two threads:
https://downloads.haskell.org/~ghc/8.4.3/docs/html/users_guide/using-concurrent.html?highlight=#rts-flag--N%20⟨x⟩.

We make it a bit more useful than that by allowing any `RTS_ARGS`.
This makes things work a little better both locally and on CI,
since locally we tend to not have weird things about cores
and can use full parallelism of our machines.