Drop `pulp` in favor of `browserify` and `purs`
`pulp` is great for when you don't want to think about a build system.
However, we've got a `Makefile`.
And our `Makefile` seems to track dependencies better than `pulp`.

What's happening in CI is that `pulp` is still taking an extraordinarily
long time on the `browserify` and `test` commands.
It looks like it's loading everything and trying to build all of it
from scratch.
We don't want this behavior.
We already know when our dependencies are up to date,
so we just use make's knowledge of the build to work more efficiently.

This seems to work our pretty alright locally.
Let's hope it works as well on CI.