Bust the cache one more time
So, here's what happened with the cache situation:
* In build 64,
    we couldn't find an appropriate `v0.1` cache for our branch.
* It did find a `v0.1` cache from build 62.
* Build 62 changed some types around.
* The output files in the `v0.1` cache from build 62 were newer than
    our source files.
* Since `purs` ultimately bases information on timestamps,
    instead of using the file contents or types or anything else,
    it thought the output files were good to use.
* The compilation failed in build 64.
* We bumped the cache version to `v0.2` and tried again.
* In build 65, we still couldn't find a cache for our branch.
* We were able to find a `v0.2` cache from build 63.
* Build 63 was on a different image,
   so it had a bunch of directories that didn't exist.
* `tar` blew up because it can't create a directory while extracting.

Basically, we had bad luck the past two builds.

The first issue could be fixed with either a different algorithm
for checking compilation artifacts (non-trivial),
or by changing the cache keys not to fallback to a sibling branch
that could be ahead of the branch we're on.

The second issue could be fixed by not setting the `working_directory`.
Circle has a default (`~/project`) that can work fine for us.
It would have prevented us needing to bust the cache to `v0.2`
in the stream of changes leading up to build 63 (which ended up causing
the second issue).

Neither solution needs to be implemented right now,
and they're both pretty orthogonal to the changes on this branch.
We're trying to bring CI times down from 17 minutes on average to
a minute or two.