PS-2961 Change the cache keys
We're hit by the second issue of:
25411248d2de374ef46da67e9ade73ea3e37d7d3.
While we could've bumped the cache version again,
there's a bigger issue here which is falling back to sibling caches.

We cannot trust a sibling branch's cache (rather than a parent branch).
We sometimes work on different things and need to change dependencies
or types or whatever.
In theory, none of this should matter.
Unfortunately, `purs` thinks that timestamps are a good way to discern
if a compiled artifact is out of date.
The punch line: timestamps are not a good way to do that.

Hopefully, this will get us where we want to be.