Remove recursion from `ExprType`
In a move to make most of our code simpler,
we trying out writing non-recursive code.

There's not much of a change yet (as `ExprType` wasn't that complex).
When we get to transformations between types,
this ought to show off its usefulness.

We're also making it easier to split apart the `ExprType`.
Splitting `ExprType` should make things a little more manageable
as the code continues to grow.
`ExprType` shouldn't be a particularly complex type,
but every bit helps to cull the growth of complexity.