Write simpler code
We're starting to use `matryoshka` more.
We can write code where we only think about what to do at one step,
and let `matryoshka` handle all of the recursive logic.

There's nothing here that's special,
nor is there anything here that couldn't have been done by hand.
The pay-off is that we don't have to do so much work.
We write code that does one thing simply, and everything works out.

Looking at the bigger picture of this commit,
we introduced a serialization layer, wrote simpler code,
and dropped a few lines of code in the process.
If that's not selling the benefit of recursion schemes,
I don't know what is.