Remove recursion from `Expr`
Again, we aim to make things simpler by removing recursion.
This time, we remove it from `Expr`.

We start to see a bigger payout in larger functions like `evalExpr`.
We don't have to explicitly remember to make recursive calls.
The data we want already exists!
We just return it.

We make a breaking change here though.
We change how `ExprType` is decoded/encoded to JSON.
Instead of having a `param :: Json` field,
its got a `params :: Array Json` field.
It's a small change, but one that unifies the two types.

It seems like this is what we want since `ExprType` being a different
type from `Expr` is an implementation detail.
Hopefully, it'll be a detail we can get rid of sooner rather than later
by separating out the cases of `Expr` and `ExprType`.