Newtype `ExprType`
Following along with the changes for  `Expr`,
we give `Mu ExprTypeF` a nominal type.
This allows us to write instances and reinforces the idea that
consumers should only really care about how to use `ExprType`
and not how `ExprType` is implemented.

Had we done this from the beginning,
there probably would've been quicker with less code written.
Live and learn.