Extract `Lookup` as its own type
We're finally breaking apart out expressions.
The separation is relatively clean here as well.
We can look at `Lookup`s behavior almost entirely in isolation from
the rest of the expression structure.

Two important changes are in `evalExprF` and the new `evalLookupF`.

`evalExprF` no longer needs a lookup function!
We only need that for `evalLookupF`.
We've simplified the implementation of `evalExprF` by detangling
what our expressions are.

This is a change only made possible by the fact that don't do anything
recursive in `evalExprF`.
If we were still doing recursion in `evalExprF`,
we'd have to thread around the lookup function and intertwine
lookups with traversing the structure.

For `evalLookupF`, we can return a pure `ExprType`.
We don't have to return some effectful `Either EvalError ExprType`!
We have relegated failure to only where it can happen.

Even though each eval function has a completely different signature,
we can compose them together and get out something that's sensible.

This ought to make some of the more complex things we have to add
(like async behavior) become just a bit easier if we can separate out
what needs to be around for synchronicity, and what doesn't.