Extract `Equal` to its own type
Fairly mechanical change for `Equal` extraction.
Again we see a benefit, namely that `Equal` is one of the things that
causes the effectful `Either EvalError ExprType` result.

If we want to focus on getting rid of the effectful result,
we should look into dealing with this case.
We don't need to even think about how `Lookup` or `Print` behaves,
because their type states that they won't fail.

Separating the `Expr` has made it a bit clearer how much work is
really necessary to make our expressions more well-typed.
We'll see at the end just how much work is actually involved though.