Extract `If` to its own type
Final bit for the main expression stuff

We see the other case making the effectful `Either EvalError ExprType`.
We can now see there are two places we'd have to deal with
if we wanted to remove effectful evaluation from our expressions.
That doesn't seem insurmountable.

Equally important,
we can see if we add something else that expands the surface area
of this effectful evaluation.
If we add another case for expressions, and it does something effectful,
we'll know instantly by the type.
We can judge whether that's the way forward or not.