Extract `ExprType` to its own module
Now that we've got all these different cases for expressions,
they each have a decent amount of behavior.
Keeping them all co-located in one module can work when there's only
a few different cases for expressions.
However, the endgame for expressions will be pretty big.
It seems to make sense that we pull them to their own modules.

The first step in that process is to pull out `ExprType`.
Ideally, `ExprType` would be just as granular as the other cases,
but we didn't do that yet.

Now that it's on its own,
it should be a little easier to split it apart as well.

Hopefully splitting out the modules makes it easier to maintain
as behavior changes and more cases are added.
If we find that it's harder to deal with,
we should pull them back into this one module.