Extract the other cases to their own modules
Here we can see just how complex each expression is (or isn't).
If we can fit the entirety of the `Lookup` behavior in 62 lines
(including boilerplate and whitespace),
I'd say we're doing pretty good considering `Lookup` is one of the
more complex types of expressions we've dealt with so far.

Let's hope this continues to be the average size for each case.

We also seem to be able to push more contextual stuff into each module.
In specific, the error handling and displaying for `If` and `Equal`
are in their respective modules.