Split `ExprType` into separate cases
Following along with the changes to `Expr`,
we can make each case of the base type separate.
We didn't separate the cases into their own modules out of lazines
and a fear of integration.
This branch is already far from `master`.
Adding more commits isn't going to help much.

The benefit here is that we ought to be able to now use these cases
to determine statically what the possible values are.

For instance, we might be able to say something like:
```PureScript
type StringRows f r
  = ( maxLength :: Inject IntF f => f
    , minLength :: Inject IntF f => f
    , placeholder :: Inject StringF f => f
    | r
    )
```

Dunno if that will work, or if that just gets us in the same position.
But the intent is that with separate cases,
we can again state what the types are statically.
And importantly, we state what they cannot be by omission.

There's still a bit more work to do,
since `Expr` and `ExprType` are both nominal types.
So we can't just drop one in for the other.
We currently have recursion in `ExprType` only for other `ExprType`s.
We'll have to think about whether we want to allow things like
`Array` to have `Expr`s within them or not.

In any case, there's still some finagling to have it work out cleanly.