PS-2968 Add `UserCleared` input
When someone is on the form, and they delete a value,
we want to know about this.
Our `UserInput` is a state machine.
And clearing the input is a valid transition for `UserInput`.

We might want to represent input more explicitly as a state machine.
But that's an idea for another day.

We'll need this state of user input when we go to add date time.
`ocelot` date time pickers can be cleared.
We have to handle that case when we get it.

We also change updating a key to deal with all user input,
instead of just setting the value.
This should make the changes for date times a bit less intrusive.