PS-2968 Add support for datetimes
This is most of the work needed to have datetime support in our forms.
There is one point we haven't finished yet:
we need to handle the datetime being unset.
The change necessary for that is available here:
https://github.com/citizennet/purescript-lynx/pull/14.

Once that PR is merged (or we somehow end up with `UserCleared`),
we can finish this up and merge this stuff in.