PS-2970 Implement a naive typeahead
This gives us the bare minimum of typeahead functionality.
We follow the pattern for dropdowns and only work with arrays of pairs.

We can continue to expand typeaheads to support the rest of the stuff
we need them to (async loading, arbitrarily structured data, etc.)
once we've fleshed out the details of what is needed for them to exist.