Add a helper for `Val`s
This is more of a preventative thing.
If we encourage using the constructor directly,
we get a bit more churn when we want to do something different with it.

We shouldn't be afraid of others using the constructor,
but we can also make their lives easier with a helper.
That helper can be a layer that reduces churn on their end.