Drop `ExprType` helpers down to `ExprType`
As we continue to add more helpers,
these few holdovers are harder to work with.
They make `Array` and `Pair` look incredibly out of place, for instance.
It also makes it hard to justify helpers for `Array` and `Pair`,
since you almost would never use them as you don't tend to want
an `Expr` (`Pair` in particular needs to be an `ExprType`).