PS-2970 Add the name into the search object
The way `fuzzy` works is that it takes an `Object String`
and searches for matches on the values of `Object`.
If we give it an object equivalent to:
```JavaScript
{ [name]: value }
```,
it won't use the `name` as a candidate for searching.

In theory, the `name` and `value` can differ.
We want to allow both as candidates for searching.

In practice, the `name` and `value` will be the same.
It shouldn't make that much of a difference.

We might think about dropping `Pair` as a thing,
if we never end up varying the `name` and `value`.