PS-2970 Parse JSON for the typeahead options
We change from having hardcoded `Expr`s to hardcoded `Json`s.
This helps us flesh out what it means to parse JSON.
We're making two similar simplifications of the problem of JSON parsing:
the JSON blob we get back ends up with a results array,
the each JSON blob in the results array ends up with a string.

We're currently assuming that there's a path from the root of
the JSON blob to the results.
Our assumption is that we get some object where the results are nested.
We don't currently handle arrays as part of this path,
but we ought to be able to make that happen by casing on the `ExprType`
in the `results` field.

Similarly, once we've got the array of results we assume there's a path
for each result down to the single field we're going to look in.

These two simplifications make parsing JSON a bit easier.
The end game is to support much more,
but we have to start somewhere.