PS-2967 Move the JSON blobs to their own file
This is preparation for loading JSON over HTTPS.
If we move the blobs to their own files,
we can reference them directly from GitHub.

It does make the feedback loop for making a change a bit longer.
But this is MVP stuff, it shouldn't be too awful to deal with.