PS-2967 Load typeahead data asynchronously
Now that we have the JSON blob available somewhere on the internet,
we can reference it and have it load up properly.
We can't do that locally without a file server running.
It seems like it ought to be possible with a new rule in the `Makefile`,
but that's a bit more work than we want to get into right now.

The URI for MVP data references GitHub's raw endpoint.
We tack an extra query key that GitHub will ignore onto the end of it.
This allows our MVP URI to match what we expect real URIs to look like.

We might want to think about supporting functions instead of
assuming the string will be the correct format.
There's no guarantee that it will be the correct format,
and it's very likely that it will lead to some hard to diagnose bugs.