PS-2967 Extract values to top level
The rendering for `TypeaheadSingle` was getting kind of out of hand.
We move stuff to the top level so it's a bit clearer what we're doing.

After moving it, it looks like these values are in the wrong module.
Neither of them really have anything to do with the component itself.
They're in this module as an artifact of the component needing them.
`itemToObject`, in particular, should live in `Lynx.Data.Expr`.