PS-2967 Move values to their respective modules
With these values in `Lynx.Page.Form`,
we were bending some abstraction boundaries.
They operate on the data in either `Lynx.Data.Expr` or `Lynx.Data.Form`.
When we need to change a value in one of those modules,
we end up hearing about it in `Lynx.Page.Form`.
We can reduce churn by putting values closer to the data they operate on.