Split apart `ExprType`
We experimented with something similar before.
Unfortunately, it was a giant PR.
To keep things maybe a little more manageable,
we try again with fewer changes.

We're going to only split apart `ExprType` this time.
This is actually where most of the money is going to be anyway.
As we should be able to use the individual types to limit values.
For instance, we could probably say an evaluated dropdown looks like:
```PureScript
type Dropdown
  = { options :: ArrayF (PairF (StringF _))
    , ...
    }
```

Once we have `Expr` in a similar situation,
we should be able to say stuff more useful:
```PureScript
type Dropdown
  = { options ::
      Free (IfF <\/> EqualF <\/> LookupF) (ArrayF (PairF (StringF _)))
    , ...
    }
```
Or something.

Let's see how it plays out once this stuff is in.

We had to move some stuff around to make this less intrusive.
In particular the `isEmpty` value moved to `Lynx.Data.Expr`.
We should try to keep values that operate on internal representations
of data near that data.
It should make future changes a bit less intrusive.