PS-2986 Split out core lynx component
The `Lynx.Page.Form` module had both the core logic of a lynx component
and the hardcoded logic of our testing and MVP views.
We want to use lynx in an internal project.
We cannot easily do that with the entangling.

We move out the core logic of a lynx component to its own module.
This ought to make it possible to use any number of these components
wherever we need to use them.