PS-2986 Only require a `PageExpr` for input
`EvalError` is completely an implementation detail of `Lynx`.
It exists entirely because our expression language is untyped.
Consumers don't need to know that,
nor do they need to be affected by our design decisions.
If we change to a typed language,
we shouldn't have to force a change on consumers because of that.

Asking only for a `Page Expr` is all we really need.
And we can make it explicit by changing the type to reflect that.