Change the way we deal with state
Before we were saying that we either had the page and its evaluated form,
or we had an error from attempting to evaluate it.

Now we're saying that we always have the page and we either have its
evaluated form or an error from attempting to evaluate it.

The difference between these is important if we want to handle failures.
In the former, we couldn't do anything once we got an error.
We'd lose the page entirely.
In the latter, we can still display the form even if there are errors.