PS-2988 Allow a page to contain tabs
We're only changing the shape of the data here.
We're not actually displaying the tabs.

This change is fairly unintrusive since we're adding another layer.
But, it would still be nice to not have to redefine every traversal.

We might consider making `Page _` a type and giving it some instances.
That might make changes a bit more cohesive.
Or, it could make things worse.
Either way, something to think about.