Ensure we always have at least one element
It doesn't make much sense for us to allow things like empty arrays.
A section with no fields is not very useful to us.

This should also make displaying tabs a bit easier.
The tab API expects there to always be some tab.
We can't guarantee that with an empty array.
If we know there's at least one tab,
we can always default in the worst case.