PS-2988 Add links to tabs
The way ocelot's navigation tabs work is by changing the route.
Each tab needs its own link so we can know which tab to show
when the route changes.

Since we don't know statically what the tab is going to be,
the best we can guess is that it'll be a `String`.
It's unfortunate, but not really much we can do about it.

We might consider changing it to `f`, so it can be an `Expr`/`ExprType`.
That change should probably come when we change the `name` similarly.