Generate fewer test cases for larger data
Because we don't control the `Arbitrary` instances
(record instances are defined elsewhere),
we were generating huge data.
This was making the tests slow.
Before this commit, we were looking at 10 minutes or so.

As a quick fix, we drop how many test cases we generate.
This makes the tests run in under 30 seconds.

An alternative would be to control the generators.
We could newtype the records and write our own instances.
We could try a different package that generated better data

In any case, this works for now.