Only evaluate expression when the fragment changes
Before, we were evaluating the new expression all the time.
Since we should be in control of the expr for a specific fragment,
we should hold on to the state and use it.

If we move fragments (i.e. we've changed pages entirely),
we should evaluate the expression, as it's likely different.