Update `ocelot` and lock down `remotedata`
As @davezuch noted in
https://github.com/citizennet/purescript-ocelot/pull/88#event-2322451142,
the `4.2.0` version of `remotedata` changed it dependencies
such that it caused a bunch of cascading failures.

Rather than continue the propagation of version failures,
we say that the maximum bound is also the minimum bound for `remotedata`.

It's not great that we have to specify an exact version,
but the alternative is worse.
Given that we're less of a library package,
and more of an spun-off application package,
it isn't that crazy to have an exact dependency anyway.