PS-2986 Shuffle modules around
In preparation for `lynx` being consumed by our internal code,
we need to make sure we don't have any conflicting modules.
We had a `Main` module that would have conflicted with our internal
`Main` module we were using elsewhere.

Taking a step back,
there's really only three modules that are `lynx`-specific:
`Lynx`, `Lynx.Data.Expr`, and `Lynx.Data.Form`.
The rest of the modules here are for putting together an example app.

To make it a bit clearer what the situation is,
we moved all of the example code to an `example` directory.

The aformentioned modules get a little shuffling themselves.
We drop the `Data` out of the middle of the latter two modules.
There is no longer a distinction between "data" and "page,"
as the "page" stuff is only used in the example.

Hopefully this makes for an easier to consume package downstream.