PS-2986 Fix name in `bower.json`
The original name of the project was `purescript-formal`.
The name as been changed for a long time.
We never had a reason to think about this though
as it hadn't been consumed by anything yet.

Now that we want to use it in an internal project,
we want the name to line up with our domain language.