Be a bit more correct with dependencies
As pointed out in
https://github.com/citizennet/orders/pull/44#discussion_r288305858,
we're not being very clear with our dependencies.
We only use `routing`/`routing-duplex` in the `examples`.
We don't use `console` at all.
And we use `quickcheck` in the `src`.

We try to make these dependencies a bit more explicit so our consumers
can have a clearer picture of what is necessary and what isn't.