Start using nix
We have a fairly simple setup,
but even it seems to have issues with different environments.

We seem to have something that's not consistent across environments,
as setting up a new sytem doesn't let `make watch` work properly.

We start using nix to track the dependencies of what we use to build
our project.

The promise of nix is that we get the exact same environment everywhere.
This is slightly different from using docker/helm/k8s in that we
only have one untracked dependency: nix.
Fortunately, nix tends to be fairly backwards compatible across versions.
It also rarely has changes in the versions since it's pretty stable.

Long and short is that this should hopefully mitigate dependency issues.