Refactor with different syntax
We can use some syntax that makes formatting a little clearer.

For instance, the `x <$ do ...` pattern ends up indenting quite a bit,
It happens because `(<$)` is not syntax but a user-land operator
and so does not have any special casing.

Things like `do` and `<-` are syntax in the language so the
source-level formatter we use (`purty`) can format them well.