Split out the source formatting from `format` rule
We want to make sure all code in the repo is formatted.
We're starting with the `src` directory and `PureScript` in particular,
but we'd like all the code to be consistently formatted.

To make it a bit easier to slot things into the build system,
we pull out the `src` directory specific parts to its own rule,
and depend on it from the `format` rule.