PS-3013 Carry more info. in `TemplateSection _`
We ran into an issue when attempting to stamp out template values.
We didn't have enough information to make the `Field _`s.
The root cause was that we were only carrying around enough information
to create `Input _`s.

We give ourselves more information to carry around for each field,
and should be able to create `Field _`s now.