PS-3014 Make a seam for `Sequence _` `Section _`s
We need to do things differently for `Sequence _` elements.
We're going to have to render them with a button to remove each element.
We want to provide some way to make that happen fairly easily.

We fork off the behavior where it will cause the least disruption.
We've copy pasted the implementation and reformatted to simplify things.
There are bigger changes coming, and it seems simpler to copy some code
than attempt to generalize when we're not sure entirely how it'll turn out.