Generate unique identifiers for sequences
We were running into the common vdom problem of a collection of children.
The long and short is that vdom-based solutions key on something
to know whether to re-render for collections of things they default to
the index of each element.

This is a common problem that affects React, elm, Halogen, etc.
There's a solution to this problem that is well documented for React:
https://reactjs.org/docs/lists-and-keys.html#keys.

We follow that solution and use a UUID as the key for our sequences.
We cannot use the index of the element,
as that's equivalent to what the vdom implementation does.
We have to use some value that uniquely identifies the value we have.
We might be able to derive this value directly from each element.
But it's unclear how we'd do that in a way that guaranteed uniqueness.

Using a UUID for this is straight forward, and solves the problem.
It does put the burden on generating unique values if someone is
creating elements for a sequence by themselves.
But that's no more a burden than what we place on the `Key`s.